Fix the accented names int Aase2012a.
The "\~{n}" sequence was meant to represent 'ñ', but instead was
breaking the name (maybe it was interpreted as a nonbreaking space?).
CensorBib was rendering "Jorge Oca\~{n}a Molinero" as
	Jorge Oca Molinero
and BibTeX (with IEEEtranS.bst) as
	J. O. na Molinero

"\'{A}lvaro D\'{i}az" was for some reason not being split into a first
and last name. CensorBib had just
	Díaz
and BibTeX had
	Álvaro Díaz
(note the non-use of a first initial in the BibTeX case; after the fix
it becomes "Á. Díaz").

Putting the braces on the outside fixes both cases.
http://www.tex.ac.uk/FAQ-bibaccent.html

Actually only "\'{A}" and "\~{n}" needed the change, but do "\'{i}" too
for consistency.

Signed-off-by: Philipp Winter <phw@nymity.ch>