Add workflows to build and deploy.
This commit adds two GitHub workflows that build and deploy CensorBib.

The "build" workflow is run on every push and verifies the validity of
the .bib file -- as seen by our build tool and *not* by LaTeX!

The "deploy" workflow is run whenenver a branch is merged into master,
at which point CensorBib is automatically built and deployed via GitHub
pages.  This means has two advantages:

* First, I no longer need to manually build and deploy CensorBib, and
  can point the domain censorbib.nymity.ch to GitHub pages.

* Second, CensorBib will also be available under
  https://nullhypothesis.github.io/censorbib/
  which may be helpful to users for who my domain is blocked.